+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `localisationService.generateDate()`
+- `localisationService.replaceSubmissionValues()`
+- `localisationService.replaceSubmissionResultValues()`
 
 ## [3.1.2] - 2023-04-20